+++ src/components/CardsManagerPage/CardsManagerModal/CardImageFinder.vue
@@ -1,15 +1,37 @@
 <template>
     <div id="card-image-find">
-        <input v-model="search" type="text" class="form-control mb-2"/>
+        <div class="input-group mb-2">
+            <span class="input-group-text">
+                <i class="fa fa-magnifying-glass"/>
+            </span>
+            <input v-model="search" type="text" class="form-control" :placeholder="$t('CardImageFinder.searchForCardImage')"/>
+        </div>
         <transition id="card-image-finder-results-container" tag="div" mode="out-in" name="translate-down-fade">
-            <div v-if="cleanSearch" id="card-image-finder-results" key="results" class="d-flex align-items-center">
-                <div v-for="imageURL of imagesURLFound" :key="imageURL">
-                    <CardImage class="mx-1" :max-width="50" :max-height="50" :image-url="imageURL" @click="selectImageURL(imageURL)"/>
-                </div>
-            </div>
-            <div v-else key="waiting-for-input" class="text-muted fst-italic small d-flex align-items-center">
+            <div v-if="!cleanSearch" key="waiting-for-input" class="text-muted fst-italic small d-flex align-items-center">
                 <i class="fa fa-chevron-up me-3"/>
                 <span v-html="$t('CardImageFinder.searchForImageByTypingRequest')"/>
+            </div>
+            <div v-else-if="isFetchingImages" key="loading" class="d-flex justify-content-center align-items-center">
+                <DefaultLoader :text="$t('CardImageFinder.loadingImages')" :icon-size="32" :text-size="20"/>
+            </div>
+            <div v-else-if="imagesFound.length" id="card-image-finder-results" key="results" class="d-flex align-items-center">
+                <div v-for="image of imagesFound" :key="image.URL">
+                    <VTooltip>
+                        <CardImage class="mx-1 cursor-pointer" :max-width="50" :max-height="50" :image-url="image.URL"
+                                   @click="selectImageURL(image.URL)"/>
+                        <template #popper>
+                            <div class="text-center" v-html="image.title"/>
+                            <hr class="my-1"/>
+                            <div class="text-center">
+                                <CardImage :max-width="100" :max-height="100" :image-url="image.URL" class="my-1 bg-white"/>
+                            </div>
+                        </template>
+                    </VTooltip>
+                </div>
+            </div>
+            <div v-else key="no-results" class="d-flex align-items-center justify-content-center text-muted">
+                <i class="fa fa-question fa-2x me-2"/>
+                <span v-html="$t('CardImageFinder.noResultFound')"/>
             </div>
         </transition>
     </div>
@@ -17,13 +39,16 @@
 
 <script>
 import Axios from "axios";
+import uniqid from "uniqid";
 import { stringify } from "qs";
 import useErrorManager from "@/composables/Error/useErrorManager";
 import CardImage from "@/components/shared/Card/Image/CardImage";
+import DefaultLoader from "@/components/shared/Loader/DefaultLoader";
+import { timeout } from "@/helpers/functions/Misc";
 
 export default {
     name: "CardImageFinder",
-    components: { CardImage },
+    components: { DefaultLoader, CardImage },
     emits: { "image-url-selected": URL => typeof URL === "string" },
     setup() {
         const { displayError } = useErrorManager();
@@ -31,8 +56,10 @@ export default {
     },
     data() {
         return {
-            imagesURLFound: [],
+            imagesFound: [],
             search: "",
+            isFetchingImages: false,
+            requestId: undefined,
         };
     },
     computed: {
@@ -49,21 +76,34 @@ export default {
     },
     methods: {
         async getImagesOnWikipedia() {
+            const requestId = uniqid();
+            this.requestId = requestId;
+            this.isFetchingImages = true;
+            await timeout(500);
+            if (requestId !== this.requestId) {
+                return;
+            }
             try {
                 const queryString = stringify({ q: this.search, limit: 20 });
                 const { data } = await Axios.get(`https://fr.wikipedia.org/w/rest.php/v1/search/title?${queryString}`);
-                this.imagesURLFound = data.pages.reduce((acc, page) => {
+                this.imagesFound = data.pages.reduce((acc, page) => {
                     if (page?.thumbnail && !acc.includes(page?.thumbnail.url)) {
-                        return [...acc, `https:${page?.thumbnail.url}`];
+                        const imageFound = { title: page.title, URL: `https:${page?.thumbnail.url}` };
+                        return [...acc, imageFound];
                     }
                     return acc;
                 }, []);
             } catch (err) {
                 this.displayError(err);
+            } finally {
+                this.isFetchingImages = false;
             }
         },
         selectImageURL(imageURL) {
             this.$emit("image-url-selected", imageURL);
+        },
+        reset() {
+            this.search = "";
         },
     },
 };